+++ features/fixtures/laravel56/app/Http/Kernel.php
@@ -56,4 +56,16 @@ class Kernel extends HttpKernel
         'hanMidEx' => \App\Http\Middleware\HandledMiddlewareEx::class,
         'hanMidErr' => \App\Http\Middleware\HandledMiddlewareErr::class,
     ];
+
+    protected function bootstrappers()
+    {
+        if (!getenv('BUGSNAG_REGISTER_OOM_BOOTSTRAPPER')) {
+            return parent::bootstrappers();
+        }
+
+        return array_merge(
+            [\Bugsnag\BugsnagLaravel\OomBootstrapper::class],
+            parent::bootstrappers(),
+        );
+    }
 }